64-bit fix.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@4804 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostDrivers/Support/win/SUPLib-win.cpp
@@ -624,7 +624,7 @@ int suplibOsIOCtl(uintptr_t uFunction, void *pvReq, size_t cbReq)
 
 
 #ifdef VBOX_WITHOUT_IDT_PATCHING
-int suplibOSIOCtlFast(unsigned uFunction)
+int suplibOSIOCtlFast(uintptr_t uFunction)
 {
     /*
      * Issue device I/O control.
